move #[cfg(test)] to lib.rs
