+++ core/merkle-mountain-range/src/lib.rs
@@ -5,13 +5,13 @@
 extern crate test;
 
 mod common;
-mod merkle_mountain_range;
 mod merkle_proof;
+mod mmr;
 
 #[allow(unused)]
 #[cfg(all(feature = "std", test))]
 mod tests;
 
 pub use common::*;
-pub use merkle_mountain_range::MerkleMountainRange;
 pub use merkle_proof::MerkleProof;
+pub use mmr::MerkleMountainRange;
